+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## ChangeLog
 
+* Added support for Transaction Traces.
+
 * Stack trace filenames have been shortened: Any thing preceding the first
   `/src/` is now removed.